server/requirements.txt
Django==4.2.11
graphene-django==3.2.1
pyjwt==2.8.0
